+++ src/pages/index.js
@@ -1,0 +1,4 @@
+export { Authentication } from "./authentication/Authentication";
+export { Login } from "./authentication/formComponents/Login";
+export { Signup } from "./authentication/formComponents/Signup";
+export { Home } from "./Home/Home";
